Scrape a Chapter of Desired Fanfiction

getChapter.absentfan Helper function of getChapter scraping a single chapter from specific entry in fanfiction.net.

getChapter(returns)

Arguments

returns
A list generated by getTitles including the fanfiction.net url and href tag of associated fanfic.

Value

The raw text scraped from a chapter of fanfiction

Examples

getChapter(list(url="https://www.fanfiction.net/s/13084808/1/Modern-Marauders",href="/s/13084808/1/Modern-Marauders"))

#>  [1] "Chapter 1- the big doors"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#>  [2] "\"Come on Prongs, you have to see this, it's unlike anything I've ever seen before. It bloody brilliant! Moony, Wormy, you guys come too!\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [3] "5th year Sirius Black yelled to his three best friends- his marauders- in the empty hallways of the infamous Potter Manor. He was running through the corridor, looking for the room that he had just found."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [4] "Right, left, left, straight, right, he thought as he was retracing his steps to the mysterious room that he just found while roaming through the halls, looking for interesting things in his new home now that his parents disowned him and kicked him out of his house. He couldn't be more glad. Except one thing- his little brother, Regulus. Their parents had been shoving ancient pureblood beliefs like that a pureblood was a higher caliber of wizard and half-bloods were meant to bow down to them while muggle-borns stole their magic and shouldn't be allowed to have a education with everyone else. Sirius doubted these things, but didn't say anything until he became a Gryffindor and met his best mates. Reg on the other hand, never had anyone like the marauders to show him that everything that their parents ever told them was false and that dark magic wasn't good. Because of this, he started hanging out with darker people like Yaxley or Avery. Walburga and Orion still had hope that Sirius would 'come to his senses' and start going into darker things. They were disappointed and angry when he told them that 'he was going to finds a nice muggle girl to shag' and they told him not to come for christmas holiday."
#>  [5] "So here he was, on December 21, 1976, looking at two giant doors that were at least 10 feet tall…. because magic. Then, Rem, Jamie, and Peter came into the room a little out of breath. They all took a moment to let the intricate designs of the ginormous doors sink in."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#>  [6] "They were wooden doors with fancy carvings all throughout the top, they could see all of the little carvings that must have taken forever to carve. This wasn't some carving spells and done, this was hand-crafted, used with muggle knives. The doors themselves were beautiful. They were made of oak wood, with brass buttons throughout. It was the most beautiful and intricate thing Sirius ever saw- and that is saying something, he grew up in a 'proper' pureblood household with ancient artifacts passed down from generations of Blacks and Crabbes."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#>  [7] "\"Wow, this is crazy pretty.\" stated Prongs, in awe."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [8] "\"Yeah, sure is,\" squeaked Wormtail, sounding afraid he said the wrong thing."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#>  [9] "\"Hey James, isn't this your dad's wing of the house- you know the place where no one was supposed to go?\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [10] "Fleamont Potter- or Flee, as Sirius called him- was an inventor and a potions master. He was famous for making a potion that smoothed hair out. Sirius would never admit this to anyone, but he used it on a daily basis. It was impossible to get anyones hair to the silky perfection without even a little hair product."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [11] "James shrugged, \"Yeah, this is probably his new invention or hobby or something like that.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [12] "\"I don't care what the bloody thing is, i just want to get a good look at it.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [13] "Sirius said with excitement. \"Okay, but Siri, be careful, you never what Dad can create.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [14] "\"Yeah, yeah, I know Jamie.\" he waved James off."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [15] "\"But, look at this. I looks all muggle-made.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [16] "The other three walked up closer to get a closer look at it. They all knew that Sirius was obsessed with all things muggle. He would often ask Remus what certain things were because he was the person with the closest muggle relative- his mother was a muggle. He was crazy about a lot of the muggle band like The Beatles, Rolling Stones, Led Zeppelin, Pink Floyd, Queen, Aerosmith, Eagles, Sex Pistols, etc. he had so many muggle rock band, muggle motorcycles, and most importantly, muggle girls in bikinis posters along with Gryffindor banners all with permanent sticking charms just to spite his parents."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [17] "Sirius went to go and look at the back of the giant doors, but there was no back to the ginormous doors, they led nowhere. He thought it was some kind of portal or sort of like something in the muggle comic books that he got remus to help smuggle into Grimmauld Place."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [18] "\"Guys, look at this. There is no back to these doors. The are like transparent like they aren't even there.\" Sirius said with excited curiosity."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [19] "\"What the bloody hell are you talking about?\" Prongs said."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [20] "\"Just look at this, you need to see it to believe it.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [21] "Remus and James rushed across the room to where Sirius was, while Peter looked at it curiously, not wanting to go any closer."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [22] "\"Bloody hell, this is amazing\" Jamie said."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [23] "\"You bet your arse it is.\" Remus replied, amazed."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [24] "\"We all are asking the question on what is on the other side of the door, so let's go and open it because we all know it is something magical.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [25] "\"Padfoot, are you sure about this? I mean it could be something dangerous,\" Peter asked wearly."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [26] "\"That just makes it even better. Watch and learn\" Sirius replied."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [27] "Peter shifted uncomfortably."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [28] "\"If you die, it's all going to be on you.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [29] "\"Nice to know you care, Jamie.\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [30] "\"Okay, enough chit chat, Padfoot, just go in already.\" Remus said, hiding his excitement."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [31] "\"Yeah, Yeah, yeah, I know, I know\" Sirius replied, but he didn't sound annoyed."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [32] "\"Ok, I can do this. I am the great Sirius Black. I am the shagger of girls. I am not afraid. I can do this.\" Sirius murmured."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [33] "\"Oh, grow a pair and open it already.\" James said in anticipation."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
#> [34] "\"Shut up Jamie, I'm going.\" Sirius said as he was creeping up to the doors."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [35] "He wrapped his trembling fingers along the brass doorknob. He wasn't shaking because he was scared (that much), but more of excitement."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> [36] "He slowly turned the knob, opening the door, and was instantly pulled into a spiralling vortex."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [37] "The three teenagers that were left in the room screamed Sirius's name."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [38] "\"Sirius!\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [39] "\"Padfoot!\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [40] "\"Siri!\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [41] "James ran through the door to follow his best friend with Remus and Peter on his tail."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [42] "A/N- So, I am not British, but American. if I did anything wrong with my British slang, please tell me"