Get information on the quantity and structure of fandom pages

This function takes the output of the getURL function and generates information based on the amount of pages and the structure of the site.

getPages(returns)

Arguments

returns
list passed from getUrl function

Value

returns a list with 4 elements a boolean that returns TRUE if there are no pages associated with the URL a string that gives the pattern of the URLs of the pages a numeric with the total number of pages a string with the url passed to the function max.entries passed from getUrl

Examples

getPages(list(url="https://www.fanfiction.net/cartoon/Invader-Zim/", max.entries=NA))

#> $noPages
#> [1] FALSE
#> 
#> $page_use
#> [1] "/cartoon/Invader-Zim/?&srt=1&r=103&p="
#> 
#> $page_num
#> [1] 11
#> 
#> $url
#> [1] "https://www.fanfiction.net/cartoon/Invader-Zim/"
#> 
#> $max.entries
#> [1] NA
#> 

getPages(list(url="https://www.fanfiction.net/game/Kingdom-Hearts/", max.entries=10))

#> $noPages
#> [1] FALSE
#> 
#> $page_use
#> [1] "/game/Kingdom-Hearts/?&srt=1&r=103&p="
#> 
#> $page_num
#> [1] 11
#> 
#> $url
#> [1] "https://www.fanfiction.net/game/Kingdom-Hearts/"
#> 
#> $max.entries
#> [1] 10
#>