Fanfiction Data Scraper

Scrapes the data on stories on Fanfiction.net by fandom.

scrapeAllEntries(story, type, max.entries = NA)

Arguments

story
Which Fandom would you like to scrape?

type
What media type is it?

max.entries
How many Entries would you like to scrape? If NA, scrape all of the entries.

Value

returns a dataframe with details about the entries scraped

Examples

scrapeAllEntries("Harry Potter", "book", max.entries=11)

#>                                      name                         author
#> 1                      55 Shades of Crazy                  Alice Brennan
#> 2  Lenka Láskorádová a Deník Temného Pána                          mcepl
#> 3                   Harry Potter’s Sister Lucy.Whisky.HamsterAnimal.Love
#> 4                                 Elected                GlimmerGamingMC
#> 5            You're A Mean One, Mr Malfoy           Mr. Cobrah Thunderer
#> 6                                 Matched                         Leca B
#> 7          Harry Potter - Three and Seven                          Neale
#> 8       where the stars hide their graves                   MorteSangriz
#> 9                                Symphony             Carolare Scarletus
#> 10                      Douleur éternelle                         Lili76
#> 11                        Falling For You                       asdnkj23
#>                                                                                                                                                                                                                                                                                                                                                                                         description
#> 1                                                                                                                                                                                                                                                                                                                             A witch raised in captivity returns to Hogwarts to avenge her mother.
#> 2                                                Plány Toma Radla se zhroutí, když Ginny Weasleyová ztratí jeho deník hned ve svém prvním roce školy a nalezne ho jistá Lenka Láskorádová. Následuje rada bizarních rozhovoru. Lenko? Ano, Tome? Hodne jsem nad tím premýšlel … a myslím, že jsi šílená. Preklad anglické povídky „Luna Lovegood and the Dark Lord's Diary" od „The madness in me".
#> 3                                                                                                                                                                                                                                                         Harry potter is rushing to the door to get to hogwarts.. but when he opens the door he sees a baby girl, who is she and why is she there?
#> 4                                                                                                             The Wizengamot does not want Rufus Scrimgeour as Minister for Magic. Who will be his replacement? And how will their decision change the future? AU Sixth year. Slightly OOC Harry. Confident!Harry, powerful!Harry. No Horcruxes. Slight Dumbledore/Ron bashing. Rated T for safety.
#> 5  Lucius Malfoy hates Christmas a whole lot, he's not going to change his mind, he's not. He has an epiphany when reading Dr. Seuss' classic character; The Grinch, he could steal Christmas instead, it'd be a cinch! This spoof, entirely in rhyme, details this entire, horrid crime. Starring Lucius Malfoy as The Grinch, Dobby as Max, and a mystery character as Cindy Lou Who, circa 1982.
#> 6                                                                                                                          Ginny stumbles through their fireplace with gleaming eyes and a look that both amuses and terrifies him. "Welcome back," Harry starts, but stops at the shit-eating grin currently on Ginny's face. "What is it this time? Theft? Murder?" "Match-making!" she exclaims.
#> 7                                                                                                                                                                                                                                                                      Another AU Harry Potter story. The theme is time travel/do-over, HP/HG/LL, AD/MW/RW/GW/VD/OD/DD/MD bashing/flaying prevalent
#> 8                         Sometimes new stars are born from the collapse of others, life from death, order from chaos. Sometimes new stars are born in worlds that once lived in the pages of a book and think upon waking "Oh, shit." Canopa is one such star. Now, if she could actually remember the entire plot instead of just bits and pieces that would be great. (SI/OC) Pairing undecided.
#> 9     Hermione Granger was sick of life stringing her along, so when an opportunity arises to enter a competition for free dance lessons, Ginny Weasley signs her up and miraculously wins. Once she learns the instructor is none other than Marcus Flint, a boy she went to school with at Hogwarts, she can't help free drawn to the over-zealous man with the impressive moves. MarcusxHermione
#> 10                                                                                                                                                                                                                                Harry a vaincu Voldemort. Pourtant, il s'isole et ne va pas bien définitivement. Hermione, inquiète fait appel à la seule personne qui pourra l'aider selon elle.
#> 11                                                                                                                                                                                                                                                                                                                        Hallie, Harry's twin, falls for Draco Malfoy when she knows she shouldn't
#>    rating language          theme chapters  words reviews favs follows updated
#> 1      K+  English  Drama/Fantasy        1    556    <NA> <NA>    <NA>      NA
#> 2       K    Czech          Humor        5  1,515    <NA> <NA>    <NA>      NA
#> 3      K+  English Fantasy/Family        2    376       1    1       4      1h
#> 4       T  English           <NA>        2  7,914       6   23      45      1h
#> 5       K  English   Parody/Humor        2  3,667       2    4       7      1h
#> 6      K+  English  Humor/Romance        1  1,768       2    3       2      NA
#> 7       T  English           <NA>        4 32,185      14   43      59      2h
#> 8       T  English           <NA>        3 10,229       7   31      54      2h
#> 9       T  English  Romance/Humor        2  4,187       5    9      10      2h
#> 10      K   French           <NA>        1  3,608       1    2       1      NA
#> 11      T  English           <NA>        1  7,029    <NA> <NA>    <NA>      NA
#>    published
#> 1        26m
#> 2         1h
#> 3         2h
#> 4  12/6/2018
#> 5  12/9/2018
#> 6         2h
#> 7         NA
#> 8  4/16/2018
#> 9   9/3/2018
#> 10        2h
#> 11        2h