Chapter to HTML Output

Generate an easily readable html document with a fully scraped fanfiction retrieved from fanfiction.net.

storyPrint(story, type, title, browse = FALSE)

Arguments

story
Name of fandom being searched. Ex, "Harry Potter."

type
Whether the fandom is from the media type of anime, book, cartoon, comic, game, misc, movie, play, or tv.

title
The title of the specific fanfic post to be scraped.

browse
If TRUE, use external browser

Value

path to HTML file

Examples

storyPrint("Harry Potter", "book", "Modern Marauders")

#> 
#> 
#> processing file: file152028a916ec.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |.............                                                    |  20%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................                                       |  40%
#> label: setup (with options) 
#> List of 1
#>  $ include: symbol F
#> 
#>   |                                                                         |.......................................                          |  60%
#>   ordinary text without R code
#> 
#>   |                                                                         |....................................................             |  80%
#> label: unnamed-chunk-1 (with options) 
#> List of 3
#>  $ warning: symbol F
#>  $ message: symbol F
#>  $ echo   : symbol F
#> 

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:absentfan':
#> 
#>     explain, intersect, setdiff, union

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#> Quitting from lines 9-21 (file152028a916ec.Rmd) 

#> Error: 'NA' does not exist in current working directory ('C:/Users/Liza/AppData/Local/Temp/Rtmpm6ONGB').

#>