Fanfiction Data Scraper

Scrapes the data on stories on Fanfiction.net by fandom.

scrapeAllEntries(story, type, max.entries = NA)

Arguments

story
Which Fandom would you like to scrape?

type
What media type is it?

max.entries
How many Entries would you like to scrape? If NA, scrape all of the entries.

Value

returns a dataframe with details about the entries scraped

Examples

scrapeAllEntries("Harry Potter", "book", max.entries=11)

#>                                               name           author
#> 1                   Is this a dream or a nightmare     PunkQueen123
#> 2                                  The Parent Trap  Misskimmypotter
#> 3                                     This is Home      09sarahbean
#> 4                    Jul-Trolldom och Galna Upptåg  Loopin' Lunan95
#> 5                    Nobody ever asked my birthday         SkyeMoor
#> 6  Siria Potter-Black and the Order of the Phoenix chamberinmyheart
#> 7                                  Ron's Last Case        KyleLandy
#> 8                  Issus de Deux Mondes Différents heroofmyownstory
#> 9                                       Fated Moon      SullysSkull
#> 10                              Fifty-Six Funerals         Chanel19
#> 11                          10 kleine Dinge, Vol 2      ChrissiTine
#>                                                                                                                                                                                                                                                                                                                                                                                      description
#> 1                                                                                                                                                                                                It's just after the war and The Golden Trio are going back to Hogwarts, but on the first night when in the empty common room four rather unexpected appears. Marauders-Jily and Ronmione fluff.
#> 2                 Ten Years after the heartbreaking divorce between Harry and Hermione, their twin daughters meet at a Summer Camp, unbeknownst to their parents. Lily and Hailey hatch a plan to reunite their parents but what happens when it's discovered Harry is engaged to another woman? Will the twin's plan come full circle or will Harry's fiancée stop their plan once and for all?
#> 3                                                                                                                                                                                                                                                 Lucy Jones, an eleven-year-old witch from America, is invited to attend Hogwarts, where she immediate finds friends and a home away from home.
#> 4                December är här och julen närmar sig i Hogwarts. Men månaden är lång och massor kan fortfarande ske, i all synnerhet för Gyllene Trio och deras hedersgrävling. Fred och George har nya upptåg på gång och Harry vet att det kommer bli en annorlunda December detta året. Speciellt när Elyon bestämmer sig att ta upp en verbal strid med Paddan. Lunan95's julkalender 2018.
#> 5                                                                                                                                                                            "Neither can live while the other survives..." Yeah, that sounds like Harry Potter, alright. Except no, it really doesn't. In this story, we choose a far more likely candidate than Harry Potter for the prophecy.
#> 6  Siria returns for her fifth year at Hogwarts, and Cassius Warrington returns for his final year. How will the two fair against the Ministry appointed Hogwarts Inquisitor? With tensions outside the castle heating up between those that believe Voldemort has returned & those that cannot, & all eyes on her, will Siria melt in the flames or rise like a phoenix? [Updates Tues & Thurs]
#> 7                                                                                                                                                                                                                    This is the story of the final mission for ex-Auror Ronald Bilius Weasley. Ron hunts down a time traveller through time as he discovers the darkest secret of Harry Potter.
#> 8                            Univers Alterné. Tenez Sirius Black, le désinvolte tombeur plein aux as de Gryffondor. Et tenez aussi Ginny Weasley, la fauchée Attrapeuse de Gryffondor, dont le seul espoir au monde est d'être engagée par une ligue majeure de Quidditch pour changer son destin. Qu'est-ce qui s'avèrera quand ces deux mondes différents s'affrontent, entrent en collision ?
#> 9                                                                                                                                                                                                                                                                                                                In 1971 Hermione Greyback, daughter of Fenrir's Greyback, is going to Hogwarts.
#> 10                                                                                                                                                                       After the Battle of Hogwarts, Harry, Ron, and Hermione try to make sense of their post-battle life and their relationships. Takes place immediately after Back at the Burrow and overlaps slightly with From the Ashes.
#> 11                                                                                                                                           Adventskalender-FF. 10 kleine Dinge über die Partner der nächsten Generation. Was beschäftigt sie? Wie haben sie sich in die Kinder der nächsten Generation verliebt, und wie kommen sie damit zurecht, Teil der berühmten Familie Weasley zu sein?
#>    rating language                theme chapters   words reviews favs follows
#> 1       T  English        Humor/Romance       26  26,536      15   14      16
#> 2       T  English       Family/Romance       10  13,849      25   47      98
#> 3      K+  English Adventure/Friendship       10  42,337      15    1      10
#> 4       T  Swedish     Friendship/Humor        1   1,831    <NA> <NA>    <NA>
#> 5      K+  English     Adventure/Family      387 247,619     295  162     214
#> 6       T  English              Fantasy       32  91,046       3   17      19
#> 7       T  English    Adventure/Mystery        2  18,788      16   16      30
#> 8       T   French              Romance        1   1,137    <NA> <NA>    <NA>
#> 9       T  English   Romance/Friendship        4  11,760      16   42      98
#> 10      T  English        Angst/Romance        5  15,133      12    4       5
#> 11      T   German       Family/Romance       13  23,097       6    4       2
#>       updated  published
#> 1  12/11/2018 11/13/2018
#> 2  11/14/2018  7/29/2018
#> 3         12m 11/12/2018
#> 4          NA        17m
#> 5  12/11/2018  3/11/2016
#> 6         22m         NA
#> 7   12/7/2018  12/4/2018
#> 8          NA        52m
#> 9          1h  9/11/2018
#> 10         1h         NA
#> 11         1h         NA