Fanfiction Data Scraper

Scrapes the data on stories on Fanfiction.net by fandom.

scrapeAllEntries(story, type, max.entries = NA)

Arguments

story
Which Fandom would you like to scrape?

type
What media type is it?

max.entries
How many Entries would you like to scrape? If NA, scrape all of the entries.

Value

returns a dataframe with details about the entries scraped

Examples

scrapeAllEntries("Harry Potter", "book", max.entries=11)

#>                                               name                author
#> 1  Siria Potter-Black and the Order of the Phoenix      chamberinmyheart
#> 2                                  Ron's Last Case             KyleLandy
#> 3                  Issus de Deux Mondes Différents      heroofmyownstory
#> 4                                       Fated Moon           SullysSkull
#> 5                               Fifty-Six Funerals              Chanel19
#> 6                           10 kleine Dinge, Vol 2           ChrissiTine
#> 7                   Partners - A Marauders What-If              Syilline
#> 8                                   HIM - wolfstar     halfravenhalfclaw
#> 9                               The Spirits' World              Mizumiii
#> 10                              Dimension Taveller              Cubismus
#> 11                                   The Pretender NeverBeyondRedemption
#>                                                                                                                                                                                                                                                                                                                                                                                      description
#> 1  Siria returns for her fifth year at Hogwarts, and Cassius Warrington returns for his final year. How will the two fair against the Ministry appointed Hogwarts Inquisitor? With tensions outside the castle heating up between those that believe Voldemort has returned & those that cannot, & all eyes on her, will Siria melt in the flames or rise like a phoenix? [Updates Tues & Thurs]
#> 2                                                                                                                                                                                                                    This is the story of the final mission for ex-Auror Ronald Bilius Weasley. Ron hunts down a time traveller through time as he discovers the darkest secret of Harry Potter.
#> 3                            Univers Alterné. Tenez Sirius Black, le désinvolte tombeur plein aux as de Gryffondor. Et tenez aussi Ginny Weasley, la fauchée Attrapeuse de Gryffondor, dont le seul espoir au monde est d'être engagée par une ligue majeure de Quidditch pour changer son destin. Qu'est-ce qui s'avèrera quand ces deux mondes différents s'affrontent, entrent en collision ?
#> 4                                                                                                                                                                                                                                                                                                                In 1971 Hermione Greyback, daughter of Fenrir's Greyback, is going to Hogwarts.
#> 5                                                                                                                                                                        After the Battle of Hogwarts, Harry, Ron, and Hermione try to make sense of their post-battle life and their relationships. Takes place immediately after Back at the Burrow and overlaps slightly with From the Ashes.
#> 6                                                                                                                                            Adventskalender-FF. 10 kleine Dinge über die Partner der nächsten Generation. Was beschäftigt sie? Wie haben sie sich in die Kinder der nächsten Generation verliebt, und wie kommen sie damit zurecht, Teil der berühmten Familie Weasley zu sein?
#> 7                                                   What if Sirius Black is forced to join the Auror Office - taking it by storm? What if Voldemort never killed James and Lily? What if the First Wizarding War is ongoing, but the tides have turned? And what if James Potter has a little sister, six years his junior, who is just beginning her carrier? This is an exercise in "What-If."
#> 8                                                                                 How do you make someone fall in love with you when they won't even talk to you? Remus Lupin is starting his last year at high school, expecting it to be normal, but when some boy shows up and make you fall madly in love, how can anything ever be normal? No, the real question is; is it the right thing?
#> 9                                                                                                                                   Dumbledore décide d'organiser un échange international entre Hogwarts et Mahoutokoro, l'école japonaise de sorciers. Deux visions totalement différentes de la magie vont alors se rencontrer et créer des étincelles qui allumeront peut-être un incendie !
#> 10               Long before Harry even knows about Hogwarts, he knows, that he has "powers". He rans away from the Dursley and survives on his own. But when he is hurt and lost, his magic fulfills him his biggest wish: a family. His family, which is still alive in a different dimension. Only that nothing is ever easy with Harry and he will need time until he trusts and is trusted.
#> 11                                                                                                                                                 Hermione was by no means an expert in Dark Magic, what she needed was someone who was. Gellert Grindelwald, one of the most powerful dark wizards of all time, was certainly no fan of this 'Voldemort'. Maybe they could reach an agreement.
#>    rating language               theme chapters  words reviews favs follows
#> 1       T  English             Fantasy       32 91,046       3   17      19
#> 2       T  English   Adventure/Mystery        2 18,788      16   16      30
#> 3       T   French             Romance        1  1,137    <NA> <NA>    <NA>
#> 4       T  English  Romance/Friendship        3  7,840      16   41      97
#> 5       T  English       Angst/Romance        5 15,133      12    4       5
#> 6       T   German      Family/Romance       13 23,097       6    4       2
#> 7       T  English   Adventure/Romance        8 15,369       4    9      11
#> 8       T  English     Romance/Tragedy       16 24,287       2 <NA>       1
#> 9       K   French     Fantasy/Romance       22 64,680      24   10      15
#> 10      T  English Family/Hurt/Comfort        9 18,097       4   15      31
#> 11      T  English           Adventure       27 40,792      73   81     133
#>       updated  published
#> 1          3m         NA
#> 2   12/7/2018  12/4/2018
#> 3          NA        33m
#> 4  11/12/2018  9/11/2018
#> 5          1h         NA
#> 6          1h         NA
#> 7          1h 11/23/2018
#> 8          1h  9/26/2018
#> 9          2h   3/7/2018
#> 10         2h         NA
#> 11         2h 11/23/2018