Fanfiction Data Scraper

Scrapes the data on stories on Fanfiction.net by fandom.

scrapeAllEntries(story, type, max.entries = NA)

Arguments

story
Which Fandom would you like to scrape?

type
What media type is it?

max.entries
How many Entries would you like to scrape? If NA, scrape all of the entries.

Value

returns a dataframe with details about the entries scraped

Examples

scrapeAllEntries("Harry Potter", "book", max.entries=11)

#>                                   name                author
#> 1                         Close To You Wingardium Leviosa 11
#> 2                      To Judge A Book           Razamataz22
#> 3              The Course of True Love          NoseyHeiffer
#> 4       Partners - A Marauders What-If              Syilline
#> 5                              Elected       GlimmerGamingMC
#> 6                         Twisted Fate         Reaper Apollo
#> 7                Contes de Noël - 2018          A.T.S Ludwig
#> 8               The Conquerer of Death             XNeon3044
#> 9           Calendrier de l'Avent 2018             Yrie Valk
#> 10 How Were James and Lily Compatible?             Carlyn858
#> 11                   Maybe, just maybe       amarauder'swife
#>                                                                                                                                                                                                                                                                                                                                                                       description
#> 1  Hermione and Draco have been happily married for three years, when suddenly Draco wakes up with no memory of their relationship or the last eight years of his life. Can they find a way to restore his memories? Can they forge a new relationship? A story about love, persistence, memory, and second chances. And maybe also about gardening. EWE, mostly canon-compliant.
#> 2                                                                                                                                                                                                                                               First impressions are a must in the magical world. It is a shame however that nobody ever takes the time to look beyond his face.
#> 3                                                                                                                                                                                                                Severus Snape must learn the difference between friendship, love, and obsession. Hermione is engaged to Ron, but is he her true love match? Or is there another?
#> 4                                    What if Sirius Black is forced to join the Auror Office - taking it by storm? What if Voldemort never killed James and Lily? What if the First Wizarding War is ongoing, but the tides have turned? And what if James Potter has a little sister, six years his junior, who is just beginning her carrier? This is an exercise in "What-If."
#> 5                                                                                           The Wizengamot does not want Rufus Scrimgeour as Minister for Magic. Who will be his replacement? And how will their decision change the future? AU Sixth year. Slightly OOC Harry. Confident!Harry, powerful!Harry. No Horcruxes. Slight Dumbledore/Ron bashing. Rated T for safety.
#> 6                                                                            Grindelwald unwittinglly made a strong impression on Dumbledore by his phrase 'For the Greater Good' and now lives are destroyed because of it. Follow Harry as he tries to help his brother defeat the Dark Lord and intervene into the manipulations of a certain headmaster. Eventual Harry/Fleur
#> 7                                                                                                                                                                                          En attendant Noël, un Calendrier de l'Avent, avec chaque jour un nouveau texte, doux comme de la laine, réconfortant comme un chocolat chaud, ou tendre comme un baiser sous le sapin.
#> 8                                                                                                                                                                   What prison could hold the One who conquered Death? This is one such take on the idea, let us hope that no one disturbs that which should be left alone to time itself. (A world-building exercise) !One-Shot
#> 9                                                                                                                                                                                                                                                                                                Qui a dit que l'Avent ne devait inspirer que des textes doux, joyeux et légers ?
#> 10                                                                                                                                                                                                                         James Potter was Black and White. Lily Evans was a blur of Light and Colour. They contrasted and met with an electric passion as fiery as Lily's hair.
#> 11           Lily and James hate each other. The thing is, they don't really know each other. So, when they're thrown in seventh year and forced to work together, they discover things about one another and themselves. ENJOY! The summary sucks but trust me, you hungry little Jily lovers, the story is way way WAY better. RATED FOR LANGUAGE AND SOME SCENES. James x Lily
#>    rating language               theme chapters  words reviews favs follows
#> 1       T  English       Romance/Drama       12 66,020      58   74     151
#> 2       T  English  Hurt/Comfort/Drama       12 42,295      88  225     345
#> 3      K+  English             Romance        4 10,099      19   15      53
#> 4       T  English   Adventure/Romance       13 30,741       7   19      27
#> 5       T  English                <NA>        3  9,747      14   49      95
#> 6       T  English Hurt/Comfort/Horror        4  3,202       7   25      29
#> 7       K   French                <NA>       20  6,407      44    3      15
#> 8       K  English                <NA>        1    553    <NA> <NA>    <NA>
#> 9       T   French        Hurt/Comfort       20  4,383       1    1       3
#> 10      K  English      Poetry/Romance        1    309    <NA> <NA>    <NA>
#> 11      T  English       Romance/Humor        2  2,142    <NA>    2       3
#>       updated  published
#> 1  11/20/2018  5/17/2018
#> 2         22m 10/23/2016
#> 3  12/17/2018  12/9/2018
#> 4         33m 11/23/2018
#> 5         44m  12/6/2018
#> 6          1h   8/1/2018
#> 7          1h  12/1/2018
#> 8          NA         1h
#> 9          1h         NA
#> 10         NA         2h
#> 11         2h  12/2/2018