Fanfiction Data Scraper

Scrapes the data on stories on Fanfiction.net by fandom.

scrapeAllEntries(story, type, max.entries = NA)

Arguments

story
Which Fandom would you like to scrape?

type
What media type is it?

max.entries
How many Entries would you like to scrape? If NA, scrape all of the entries.

Value

returns a dataframe with details about the entries scraped

Examples

scrapeAllEntries("Harry Potter", "book", max.entries=11)

#>                                                 name              author
#> 1                                    Rewriting Stars               VSnow
#> 2                        The Problems of Time Travel             ctc1000
#> 3                                  Rewriting Destiny        mayawrites95
#> 4                              Savicorns & Fireflies   Rainbow-Pages0618
#> 5                             Kiss under a Full Moon            LexSnape
#> 6  Merry Xmas everybody - a Tedtoire Christmas story   EmbraceTheMystery
#> 7                           Wonderful Christmas Time      SerdaiglePower
#> 8                                    A Game of Chess             Kirinin
#> 9                              Between Like and Love nottheonlyfanaround
#> 10                                  So Nice So Smart          valyrian23
#> 11                                         The ideal      PhoenixKrieger
#>                                                                                                                                                                                                                                                                                                                                                                                        description
#> 1                                                                                                                                                                                                                                                     The story of Albus and Gellert's meeting, relationship and parting. In some ways they were together even after, until death would they part.
#> 2                                                              A massive change to the time line creates a world where Harry Potter attends a different Hogwarts. When he meets Hermione Granger again, dark clouds gather, and Harry has to deal with the challenges of being a hero. Harry has a clue, but his enemies are deadlier than canon. Year 4. Some action, some politics, some banter.
#> 3  They thought after Voldemort's fall that the world would get better. But they were wrong. The Death Eaters are still ruling, using politics instead of hexes. With the dwindling Wizarding population and no one left to fight, they must go back in time to before Voldemort rose to power to fix what broke in the past. Nominated for Best James and Best Peter in the 2018 Marauder Medals!
#> 4                                                         A new creature appears at Malfoy Manor, causing the curious to come looking. Draco must seek help from the Head of the Department for Regulation and Control of Magical Creatures: Hermione Granger. In order to save the creature and find its place in the Wizarding World, they must overcome the past and help one another. Dramione
#> 5                                                                                                                                                                                                                                                                                                        Harry mira la luna llena, mientras espera a que Fenrir regrese a casa. [Kissing list #32]
#> 6                            It's Victoire Weasley's final year at Hogwarts. She has to take her N.E.W.T.s, figure out what she wants to do after school and get over her crush on her childhood best friend, Teddy Lupin. When the time comes to celebrate Christmas at the Burrow, Victoire is more anxious than ever. A new chapter will be uploaded every day until Christmas! T for language.
#> 7                                                                                                                                                                                                                                                                                           Calendrier de l'Avent 2018. Un personnage et un texte par jour sur le thème de Noël et d'Harry Potter.
#> 8                    The war is all but over, and the wrong side won. Worse, the Wizarding World's hero and Ron Weasley's best friend died in the fight. When Draco Malfoy offers Ron a way to go back and fix things, he jumps at the chance. But can he anticipate how his changed moves will affect the board? How long can a king masquerade as a pawn? [Trope stew: read the authors' notes.]
#> 9                                                                                                                                                                                                                                                                                          A GinnyxDraco story in which Ginny reminisces on the difference between liking someone and loving them.
#> 10                 They say some of the greatest things that happen to us are the things we didn't plan for. Whoever said that was right, I had no idea those few steps I took started a chain of events that led me to be at the right place at the right time. The only problem with this is the right place and the right time may happen, the rest is a gamble you have to be willing to take.
#> 11                                                                                                                                                                                                                                                                                                                                                                                       The idea.
#>    rating language                theme chapters   words reviews favs follows
#> 1       T  English              Romance       10  14,297       5    1       3
#> 2       T  English            Adventure       25  96,872      92   80     189
#> 3       T  English    Adventure/Romance       56 235,598     787  448     796
#> 4       T  English Hurt/Comfort/Fantasy       10  16,527      20   22      58
#> 5       K  Spanish              Romance        1     716    <NA> <NA>    <NA>
#> 6       T  English        Romance/Drama        6   4,609       2    1       1
#> 7       K   French               Family       19  15,752      18    1       2
#> 8       T  English        Drama/Mystery        4  21,181      20   10      27
#> 9       T  English                 <NA>        1   1,219    <NA> <NA>    <NA>
#> 10     K+  English Hurt/Comfort/Romance        1   1,740    <NA> <NA>    <NA>
#> 11      K  English                Humor        1       2    <NA> <NA>    <NA>
#>       updated  published
#> 1  12/17/2018  12/4/2018
#> 2          4m 10/31/2018
#> 3  12/16/2018   1/1/2018
#> 4          1h   8/6/2018
#> 5          NA         NA
#> 6  12/17/2018 12/13/2018
#> 7         40m  12/1/2018
#> 8         46m         NA
#> 9          NA         1h
#> 10         NA         1h
#> 11         NA         1h